Archive/1.1/requirements.txt
psutil
pypresence
pystray
Pillow
flask
tkinter
